+++ src/Euterpe.Core/Extensions/CoreServiceExtensions.cs
@@ -1,4 +1,5 @@
 using Euterpe.Core.Http.Handlers;
+using Euterpe.Core.Http.Listeners;
 using Microsoft.Extensions.Http.Resilience;
 using Polly;
 using SoundFlow.Abstracts;
